refactor: decouple panels from axes 01 - moved the overall panel logic into panels
